organize styles and add barrel exports
