+++ client/src/App.js
@@ -1,23 +1,8 @@
 import React from "react";
 import { ThemeProvider } from "./theme/ThemeProvider";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Navbar from "./components/Navbar";
-import Home from "./pages/Home";
-import Syllabus from "./pages/Syllabus";
-import Notes from "./pages/Notes";
-import PYQs from "./pages/PYQs";
-import Feedback from "./pages/Feedback";
-import About from "./pages/About";
-import Privacy from "./pages/Privacy";
-import Contact from "./pages/Contact";
-import Analytics from "./pages/Analytics";
-import Faq from "./pages/Faq";
-import Contribute from "./pages/Contribute";
-import ScrollToTop from "./components/scrolltotop";
-import MindMapEditor from "./pages/MindMapEditor";
-import Profile from "./pages/Profile.js";
-import SubmitFeedback from "./pages/SubmitFeedback";
-import Todo from "./pages/Todo";
+import {Home, Syllabus, Notes, PYQs, Feedback, About, Privacy, Contact, Analytics, Faq, Contribute, MindMapEditor, Profile, SubmitFeedback, Todo} from './pages';
+import {Navbar, ScrollToTop} from "./components";
 
 
 const App = () => {
