+++ client/src/components/index.js
@@ -1,0 +1,4 @@
+export { default as Calender} from "./Calender.js";
+export { default as FeedbackModal} from "./FeedbackModal.js";
+export { default as Navbar} from "./Navbar.js";
+export { default as scrolltotop} from "./scrolltotop.js";
