+++ client/src/pages/Contact.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import './Contact.css';
-import './ScrollToTop.css';
+import './style/Contact.css';
+import './style/ScrollToTop.css';
 
 const Contact = () => {
