+++ client/src/pages/Home.js
@@ -5,8 +5,8 @@ import { motion, AnimatePresence } from 'framer-motion';
 import { FaGithub, FaLinkedin, FaDiscord, FaArrowUp } from "react-icons/fa";
 import { FaXTwitter } from "react-icons/fa6";
 import logo from "./logo.png";
-import "./Home.css";
-import './ScrollToTop.css';
+import "./style/Home.css";
+import './style/ScrollToTop.css';
 
 const Home = () => {
