+++ client/src/pages/Notes.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useMemo, useEffect, useRef } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { FaArrowUp } from "react-icons/fa";
-import './Notes.css';
+import './style/Notes.css';
 
 const Notes = () => {
