+++ client/src/pages/style/Notes.css
