+++ client/src/pages/style/Privacy.css
